+++ lib/wrappers/spectate.dart
@@ -1,0 +1,15 @@
+import 'package:objd/basic/command.dart';
+import 'package:objd/basic/entity.dart';
+import 'package:objd/basic/rest_action.dart';
+import 'package:objd/build/context.dart';
+
+class Spectate extends RestActionAble {
+  final Entity target;
+
+  Spectate(this.target);
+
+  @override
+  generate(Context context) {
+    return Command('spectate ' + target.toString());
+  }
+}